refactor(contracts): keep group placement with sessions
