+++ packages/contracts/src/observer.ts
@@ -1,7 +1,9 @@
 import { z } from "zod";
 import type { CommandReceipt, CommandRecord } from "./commandLifecycle.js";
-import { SessionFreshStartConsentSchema } from "./commands/session.js";
-import { FreshSessionGroupPlacementIntentSchema } from "./commands/sessionGroup.js";
+import {
+  FreshSessionGroupPlacementIntentSchema,
+  SessionFreshStartConsentSchema,
+} from "./commands/session.js";
 import { RemoveWorktreePayloadSchema } from "./commands/worktree.js";
 import type { StationCommand } from "./commands.js";
 import type {
